requirements/prod.txt
mechanicalsoup==0.11.0
requests==2.20.1
urllib3==1.24.1
keyring==16.0.2
lxml==4.2.5
lz4==2.1.2
pyexcel==0.5.9.1
pyexcel-io==0.5.9.1
pyexcel-ods3==0.5.2
pyexcel-webio==0.1.4
pyexcel-xls==0.5.8
pyexcel-xlsx==0.5.6
cookiecutter==1.6.0
w3lib==1.19.0
pycryptodome==3.7.0
gevent==1.3.7
newt.db==0.9.0
zodbpickle==1.0.2
persistent==4.4.3
zodb==5.5.1